[S180928_001] Retour de tests : Erreur sur rule321 Integer instead Float

+++ webavs-web/src/main/java/globaz/apg/rapg/rules/Rule321.java
@@ -82,7 +82,7 @@ public class Rule321 extends Rule {
         }
         
         if (serviceType.equals(APGenreServiceAPG.Maternite.getCodePourAnnonce())) {
-            if ((Integer.valueOf(basicDailyAmount) > montantMiniAllocationJournalierMaternite)) {
+            if ((Float.valueOf(basicDailyAmount) > montantMiniAllocationJournalierMaternite)) {
                 return false;
             }
         } else {
